addressing comments/error handling/stylecop issues
