+++ packages/Microsoft.Bot.Components.Telephony/Microsoft.Bot.Components.Telephony.csproj
@@ -1,34 +1,29 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Version Condition=" '$(PreviewPackageVersion)' == '' ">$(LocalPackageVersion)</Version>
-    <Version Condition=" '$(PreviewPackageVersion)' != '' ">$(PreviewPackageVersion)</Version>
-    <PackageVersion Condition=" '$(PreviewPackageVersion)' == '' ">$(LocalPackageVersion)</PackageVersion>
-    <PackageVersion Condition=" '$(PreviewPackageVersion)' != '' ">$(PreviewPackageVersion)</PackageVersion>
-    <Configurations>Debug;Release</Configurations>
-    <DocumentationFile>bin\$(Configuration)\netstandard2.0\Microsoft.Bot.Builder.Dialogs.Adaptive.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup>
+    <OutputType>Library</OutputType>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <PackageId>Microsoft.Bot.Components.Telephony</PackageId>
-    <Description>This library implements .NET support for adaptive dialogs with Telephony</Description>
-    <Summary>This library implements .NET support for adaptive dialogs with Microsoft Telephony</Summary>
-    <ContentTargetFolders>content</ContentTargetFolders>
-    <PackageTags>msbot-action;msbot-trigger;msbot-component</PackageTags>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DebugType>Full</DebugType>
-    <DebugSymbols>true</DebugSymbols>
   </PropertyGroup>
 
   <PropertyGroup>
-    <!-- These documentation warnings excludes should be removed as part of https://github.com/microsoft/botbuilder-dotnet/issues/4052 -->
-    <NoWarn>$(NoWarn);CS1591</NoWarn>
-    <Version>1.0.0-preview</Version>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageId>Microsoft.Bot.Components.Telephony</PackageId>
+    <VersionPrefix>1.0.0-preview</VersionPrefix>
+    <Description>This library implements .NET support for adaptive dialogs with Telephony.</Description>
+    <Summary>This library implements .NET support for adaptive dialogs with Telephony.</Summary>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>..\..\build\35MSSharedLib1024.snk</AssemblyOriginatorKeyFile>
+    <DelaySign>true</DelaySign>
+    <ContentTargetFolders>content</ContentTargetFolders>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Bot.Builder.Dialogs.Adaptive.Runtime" Version="4.13.0" />
+
+    <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.333">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
 
   <ItemGroup>
     <Content Include="**/*.dialog" />
@@ -39,12 +34,10 @@
     <Content Include="**/*.qna" />
   </ItemGroup>
 
-  <ItemGroup>
-    <None Remove="Schemas\CallTransferDialog.schema" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Bot.Builder.Dialogs.Adaptive" Version="4.13.0-rc4" />
-  </ItemGroup>
-
+  <PropertyGroup>
+    <!-- Disable warning for SA0001 "XML comment analysis is disabled due to project configuration" which is not true -->
+    <!-- Disable warning for SA1649 "file name should match first type name" due to use of generics -->
+    <NoWarn>$(NoWarn),SA0001,SA1649</NoWarn>
+  </PropertyGroup>
+  
 </Project>
