+++ packages/Microsoft.Bot.Components.Telephony/Schemas/Microsoft.Telephony.CallTransfer.schema
@@ -4,12 +4,15 @@
   "title": "Call Transfer",
   "description": "Phone number to transfer the call to (in E.164 format such as +1425123456)",
   "type": "object",
+  "required": [
+    "phoneNumber"
+  ],
   "additionalProperties": false,
   "properties": {
-    "targetPhoneNumber": {
+    "phoneNumber": {
       "$ref": "schema:#/definitions/stringExpression",
-      "title": "Target Phone Number",
-      "description": "Phone number to transfer the call to",
+      "title": "Phone Number",
+      "description": "Phone number to transfer the call to.",
       "examples": [
         "in E.164 format such as +1425123456"
       ]
